Optifine 3

+++ src/main/java/net/optifine/ChunkPosComparator.java
@@ -31,7 +31,7 @@ public class ChunkPosComparator implements Comparator<ChunkCoordIntPair>
         int i = cp.chunkXPos - this.chunkPosX;
         int j = cp.chunkZPos - this.chunkPosZ;
         int k = i * i + j * j;
-        double d0 = MathHelper.func_181159_b(j, i);
+        double d0 = MathHelper.func_181159_b((double)j, (double)i);
         double d1 = Math.abs(d0 - this.yawRad);
 
         if (d1 > Math.PI)
@@ -39,7 +39,7 @@ public class ChunkPosComparator implements Comparator<ChunkCoordIntPair>
             d1 = (Math.PI * 2D) - d1;
         }
 
-        k = (int)(k * 1000.0D * this.pitchNorm * d1 * d1);
+        k = (int)((double)k * 1000.0D * this.pitchNorm * d1 * d1);
         return k;
     }
 }
